Separate class and subject related functions
