+++ backend/src/algorithm/utils/subject.ts
@@ -1,0 +1,5 @@
+import { Subject } from 'algorithm/types';
+
+export function getTotalTimesPerWeek({ timesPerWeek }: Subject) {
+  return timesPerWeek instanceof Array ? timesPerWeek.reduce((a, b) => a + b, 0) : timesPerWeek;
+}
